1.完成 周边景点搜索 2.增加luckmoney(红包)接口，未实现逻辑

+++ server/common/src/main/java/com/itravel/server/interfaces/dal/IAttractions.java
@@ -2,9 +2,9 @@ package com.itravel.server.interfaces.dal;
 
 public interface IAttractions {
 
-	public abstract int getId();
+	public abstract long getId();
 
-	public abstract void setId(int id);
+	public abstract void setId(long id);
 
 	public abstract String getAddress();
